Add test for non-blocking I/O
This change moves to using ssize_t for consistently tracking the
size of user data, and adds a test case around non-blocking I/O
and how s2n handles the EAGAIN/EWOULDBLOCK cases.

+++ tls/s2n_connection.h
@@ -98,7 +98,7 @@ struct s2n_connection {
     /* How much of the current user buffer have we already
      * encrypted and have pending for the wire.
      */
-    uint32_t current_user_data_consumed;
+    ssize_t current_user_data_consumed;
 
     /* An alert may be fragmented across multiple records,
      * this stuffer is used to re-assemble.
